+++ openerp/report/__init__.py
@@ -31,5 +31,14 @@ import report_sxw
 import printscreen
 
 
+def render_report(cr, uid, ids, name, data, context=None):
+    """
+    Helper to call ``ir.actions.report.xml.render_report()``.
+    """
+    import openerp
+    registry = openerp.modules.registry.RegistryManager.get(cr.dbname)
+    return registry['ir.actions.report.xml'].render_report(cr, uid, ids, name, data, context)
+
+
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4: