+++ openerp/tools/test_reports.py
@@ -26,6 +26,7 @@
 """
 
 import openerp.netsvc as netsvc
+import openerp.report
 import openerp.tools as tools
 import logging
 import openerp.pooler as pooler
@@ -50,7 +51,7 @@ def try_report(cr, uid, rname, ids, data=None, context=None, our_module=None):
     else:
         rname_s = rname
     _logger.log(netsvc.logging.TEST, "  - Trying %s.create(%r)", rname, ids)
-    res = netsvc.LocalService(rname, cursor=cr).create(cr, uid, ids, data, context)
+    res = openerp.report.render_report(cr, uid, ids, rname_s, data, context)
     if not isinstance(res, tuple):
         raise RuntimeError("Result of %s.create() should be a (data,format) tuple, now it is a %s" % \
                                 (rname, type(res)))