+++ doc/changelog.rst
@@ -6,6 +6,8 @@ Changelog
 `trunk`
 -------
 
+- Almost removed ``LocalService()``. For reports,
+  ``openerp.report.render_report()`` can be used.
 - Added the :ref:`Long polling <longpolling-worker>` worker type.
 - Added :ref:`orm-workflows` to the ORM.
 - Added :ref:`routing-decorators` to the RPC and WSGI stack.